+++ CHANGES.rst
@@ -43,6 +43,20 @@ New features:
 
 Bug fixes:
 
+- Install plone.app.theming in 5.0 alpha.
+  When it is already installed, upgrade it.
+  [maurits]
+
+- Fixed AttributeError ``use_content_negotiation`` when migrating old language tool.
+  Not all versions have the same properties available.
+  Now we only take over existing properties.
+  5.0 beta.
+  [maurits]
+
+- Fixed ConstraintNotSatisfied when default_editor is not allowed.
+  5.0 alpha.
+  [maurits]
+
 - Enabled update from latest 4.3 profile revision.
   Otherwise we would skip a few upgrade steps when migrating to
   Plone 5.  [maurits]